dissolve StretchTiling into LayerCache and streamline the documentation
StretchTiling had utility class semantics without a name saying so, and
its routines were used exclusively by LayerCache - which is small enough
to absorb them. The split follows data ownership:

- Key derivation (_canonicalize, _isStretchTileable, _sliceInsets,
  _canonicalSize, _isBlitCompatible) becomes private statics on
  LayerCache, since it exists to compute the cache key and gate the
  paint path. All of it is private now - the specs observe behavior
  through the public API and no longer need package visibility.

- Pixel reconstruction moves onto CachedImage, which owns the exemplar
  image and its stretch tiles: paintStretched() lazily extracts the
  dedicated tiles and reassembles any actual size, replacing the
  getStretchTiles/initStretchTiles accessor pair and the free-floating
  blit/extractStretchTiles functions. _copyRegion now reuses
  CachedImage._allocate instead of duplicating the compatible-image
  allocation, and the slice insets are computed once per blit instead
  of twice.

The documentation is condensed to one high level story in the LayerCache
class javadoc (exemplar-as-texture-atlas, the homogeneity invariant, the
render-input vs cache-key distinction), with method docs reduced to the
constraints the code cannot show - notably keeping the XRender interior
sub-rectangle warning on _extractStretchTiles and the device space
snapping rationale on paintStretched. Net -190 lines, no behavior change,
full suite green.
